+++ vike/shared/hooks/getHook.ts
@@ -1,18 +1,23 @@
 export { getHook }
 export { assertHook }
 export { assertHookFn }
+export { getHookTimeouts }
 export type { Hook }
 export type { HookName }
 export type { HookLoc }
+export type { HookTimeouts }
 
 import { PageContextExports } from '../getPageFiles.js'
-import type { ConfigBuiltIn, HookName } from '../page-configs/Config.js'
+import type { ConfigBuiltIn, HookName, ConfigHooksTimeouts } from '../page-configs/Config.js'
 import { assert, assertUsage, checkType, isCallable } from '../utils.js'
-import { getHookTimeouts, type HookTimeouts } from './getHookTimeouts.js'
 
 type Hook = HookLoc & { hookFn: HookFn; hookTimeouts: HookTimeouts }
 type HookLoc = { hookName: HookName; hookFilePath: string }
 type HookFn = (arg: unknown) => unknown
+type HookTimeouts = {
+  timeoutErr: number
+  timeoutWarn: number
+}
 
 function getHook(pageContext: { config: ConfigBuiltIn } & PageContextExports, hookName: HookName): null | Hook {
   if (!(hookName in pageContext.exports)) {
@@ -46,4 +51,36 @@ function assertHookFn(
   assert(!hookName.endsWith(')'))
   assertUsage(isCallable(hookFn), `Hook ${hookName}() defined by ${hookFilePath} should be a function`)
   checkType<HookFn>(hookFn)
+}
+
+function getDefaultTimeouts(hookName: HookName): HookTimeouts {
+  if (hookName === 'onBeforeRoute') {
+    return {
+      timeoutErr: 5 * 1000,
+      timeoutWarn: 1 * 1000
+    }
+  }
+  if (hookName === 'onBeforePrerender') {
+    return {
+      timeoutErr: 10 * 60 * 1000,
+      timeoutWarn: 30 * 1000
+    }
+  }
+  return {
+    timeoutErr: 30 * 1000,
+    timeoutWarn: 4 * 1000
+  }
+}
+
+function getHookTimeouts(configHooksTimeouts: ConfigHooksTimeouts | undefined, hookName: HookName): HookTimeouts {
+  const defaultHookTimeouts = getDefaultTimeouts(hookName)
+  if (!configHooksTimeouts || !(hookName in configHooksTimeouts)) {
+    return defaultHookTimeouts
+  }
+  const timeoutErr = configHooksTimeouts[hookName]?.error || defaultHookTimeouts.timeoutErr
+  const timeoutWarn = configHooksTimeouts[hookName]?.warning || defaultHookTimeouts.timeoutWarn
+  return {
+    timeoutErr,
+    timeoutWarn
+  }
 }
